move XDSClient interface definition to client.go
